Ganti icon

+++ app/src/main/java/com/example/aldy/difacademy/Activity/QuizActivity.java
@@ -28,13 +28,9 @@ import com.example.aldy.difacademy.Notification.MyResponse;
 import com.example.aldy.difacademy.Notification.Sender;
 import com.example.aldy.difacademy.Notification.Token;
 import com.example.aldy.difacademy.R;
-import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.android.gms.tasks.Task;
 import com.google.firebase.Timestamp;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
@@ -42,8 +38,6 @@ import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
 import com.google.firebase.firestore.Source;
-import com.google.firebase.iid.FirebaseInstanceId;
-import com.google.firebase.iid.InstanceIdResult;
 
 import java.util.ArrayList;
 
@@ -309,7 +303,7 @@ public class QuizActivity extends AppCompatActivity {
             }
         }
 
-        nilai = (int) (benar / totalNomor) * 100;
+        nilai = (benar / totalNomor) * 100;
 
         String message = "";
         String title = "";
